+++ code/registers.h
@@ -5,6 +5,6 @@ int32_t REGISTERS[REGISTERS_NUMBERS];
 int32_t PC, HI, LO;
 
 void initRegisters(void);
-int32_t readRegister(int index);
-int writeRegister(int index, int32_t value);
+int32_t readRegister(int8_t index);
+int writeRegister(int8_t index, int32_t value);
 void printRegisters(void);
